chore: renaming vitest to test
