+++ src/index.js
@@ -2,5 +2,5 @@ module.exports = {
   base: require("./base.js"),
   react: require("./react.js"),
   typescript: require("./typescript.js"),
-  vitest: require("./vitest.js"),
+  test: require("./test.js"),
 };
